Update example-react-modal.mdx
Previous version of this file what seemed to use an older syntax of `testing-library`. It involved direct DOM manipulation and use of ReactDOM. I corrected it to what I believe is the more modern way to test a Modal in `testing-library`. If I have something wrong, please let me know. Thanks!